ci: Renamed package to exclude conflicts
