+++ src/main/java/bjs2_57133/Calculator.java
@@ -1,4 +1,4 @@
-package school.faang;
+package bjs2_57133;
 
 public interface Calculator<T> {
     T calculate(T x, T y);
